pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "greycloud"
version = "1.0.0"
description = "Reusable Google GenAI/Vertex AI Client Module"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "GreyCloud Contributors"}
]
keywords = ["google", "vertex-ai", "genai", "gemini", "ai", "ml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "google-genai>=0.2.0",
    "google-auth>=2.0.0",
]

[project.optional-dependencies]
storage = [
    "google-cloud-storage>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "coverage>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "greycloud[storage,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/greycloud"
Documentation = "https://github.com/yourusername/greycloud#readme"
Repository = "https://github.com/yourusername/greycloud"
Issues = "https://github.com/yourusername/greycloud/issues"

[tool.setuptools]
packages = ["greycloud"]

[tool.setuptools.package-data]
greycloud = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "auth: Authentication tests",
    "batch: Batch processing tests",
]

[tool.coverage.run]
source = ["greycloud"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
